more refinments to try and get IoT working.  Still WIP.

+++ src/bitcoin/secp256k1/src/scalar_4x64_impl.h
@@ -540,7 +540,7 @@ static void secp256k1_scalar_reduce_512(secp256k1_scalar *r, const uint64_t *l) 
     sumadd_fast(n3);
     extract_fast(m5);
     VERIFY_CHECK(c0 <= 1);
-    m6 = c0;
+    m6 = static_cast<uint32_t>(c0);
 
     /* Reduce 385 bits into 258. */
     /* p[0..4] = m[0..3] + m[4..6] * SECP256K1_N_C. */
@@ -560,7 +560,7 @@ static void secp256k1_scalar_reduce_512(secp256k1_scalar *r, const uint64_t *l) 
     muladd_fast(m6, SECP256K1_N_C_1);
     sumadd_fast(m5);
     extract_fast(p3);
-    p4 = c0 + m6;
+    p4 = static_cast<uint32_t>(c0 + m6);
     VERIFY_CHECK(p4 <= 2);
 
     /* Reduce 258 bits into 256. */
